Directions button is working again
