+++ src/components/screens/TrailScreen.js
@@ -1,9 +1,6 @@
 import React, {
   Component,
 } from "react";
-import {
-  Linking,
-} from "react-native";
 import PropTypes from "prop-types";
 import {
   bindActionCreators,
@@ -12,16 +9,12 @@ import {
   connect,
 } from "react-redux";
 import * as subLocationActions from "../../actions/subLoactionActions";
-import {
-  LocationUtils,
-} from "../../utils/";
 import MapWithListView from "../shared/MapWithListView";
 
 class TrailScreen extends Component {
   static propTypes = {
     navigation: PropTypes.object, // eslint-disable-line react/require-default-props
     subLocation: PropTypes.object.isRequired,
-    geolocation: PropTypes.any.isRequired,
   }
 
   static navigationOptions = ({ navigation }) => {
@@ -57,55 +50,18 @@ class TrailScreen extends Component {
     return trailObjects;
   }
 
-  static async openUrlIfValid(url) {
-    try {
-      const supported = await Linking.canOpenURL(url);
-      if (supported) {
-        return Linking.openURL(url);
-      }
-    } catch (error) {
-      return null;
-    }
-    return null;
-  }
-
   constructor(props) {
     super(props);
     this.state = {
-      geolocation: this.props.geolocation,
       subLocation: this.props.subLocation,
       trailObjects: this.props.trailObjects,
     };
-  }
-
-  onMarkerPressed = (marker) => {
-    this.scrollToListItemWithId(marker);
-  }
-
-  scrollToListItemWithId = (marker) => {
-    const index = this.state.trailObjects.findIndex(item => item === marker);
-    this.scrollToIndex(index);
-  }
-
-  onListItemDirectionsButtonPressed = (listItem) => {
-    const locationItem = this.locationItemFromId(listItem.item.locationId);
-    const { latitude, longitude } = locationItem;
-    const directionsUrl = LocationUtils.directionsUrl(latitude, longitude, this.state.geolocation);
-    TrailScreen.openUrlIfValid(directionsUrl);
   }
 
   locationItemFromId = (locationId) => {
     const { _embedded } = this.state.subLocation;
     const locationItem = _embedded.location.filter(item => item.id === locationId);
     return locationItem[0];
-  }
-
-  getDistancefromUserLocationToLocationItem(locationItem) {
-    if (!this.state.geolocation) return 0;
-
-    const { coords } = this.state.geolocation;
-    const distance = LocationUtils.getDistanceBetweenCoordinates(locationItem, coords);
-    return distance;
   }
 
   render() {
